Fixed a problem with a guest user trying to add annotations
