--- app/javascript/packs/dictionary.js
@@ -1,12 +1,0 @@
-// Run this example by adding <%= javascript_pack_tag 'hello_react' %> to the head of your layout file,
-// like app/views/layouts/application.html.erb. All it does is render <div>Hello React</div> at the bottom
-// of the page.
-
-import React from 'react';
-import ReactDOM from 'react-dom';
-import PropTypes from 'prop-types';
-import Dictionary from '../widgets/Dictionary/Dictionary';
-
-document.addEventListener('DOMContentLoaded', () => {
-  !!document.getElementById("Dictionary") && ReactDOM.render(<Dictionary />, document.getElementById('Dictionary'));
-})
