+++ app/views/authors/by_section.html.haml
@@ -1,5 +1,3 @@
 .c
   %h1= @section.identifier
   #Highlight{ 'data-author' => @author.id, 'data-text' => @text.id, 'data-book' => @book.id, 'data-section' => @section.id }
-
-= javascript_pack_tag 'highlight'
