+++ app/views/layouts/application.html.erb
@@ -16,6 +16,6 @@
     <%= render "partials/navigation" %>
     <%= render "partials/dictionary" %>
     <%= yield %>
-    <%= javascript_pack_tag 'dictionary' %>
+    <%= javascript_pack_tag 'application' %>
   </body>
 </html>
